pyproject.toml
[tool.poetry]
name = "leaderboard-github-bot"
version = "0.1.0"
description = ""
authors = ["Artur A. Galstyan <galstyan.artu@gmail.com>"]
readme = "README.md"
packages = [{include = "leaderboard_github_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "leaderboard_github_bot.main:dev"
start = "leaderboard_github_bot.main:start"
